% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplot.R
\name{mvHaplotype}
\alias{mvHaplotype}
\title{Redirect haPLOType app.}
\usage{
mvHaplotype(path)
}
\arguments{
\item{path}{string. directory path. Required}
}
\description{
makes a copy of shiny haPLOType to a different directory
}
\examples{
\dontrun{
mvHaplotype("~/Shiny")
}
}
