% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplot.R
\name{mvShinyHaplot}
\alias{mvShinyHaplot}
\title{Transfer a copy of microHaplot app.}
\usage{
mvShinyHaplot(path)
}
\arguments{
\item{path}{string. directory path. Required}
}
\description{
Moves shiny haPLOType app to a different directory
}
\examples{
\dontrun{
mvShinyHaplot(tempdir())
}
}
