% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplot.R
\name{runHaplotype}
\alias{runHaplotype}
\title{Run Haplotype Plot.}
\usage{
runHaplotype(path = system.file("shiny", "microhaplot", package =
  "microhaplot"))
}
\arguments{
\item{path}{Path to shiny haPLOType app. Optional. If not specified, the path is default to local app path.}
}
\description{
Run Haplotype shiny
}
\examples{
# this starts a shiny session, so should not be run during R CMD CHECK, etc.
\dontrun{
runHaplotype()
}
}
