% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplot.R
\name{runShinyHaplot}
\alias{runShinyHaplot}
\title{Run shiny microhaplot}
\usage{
runShinyHaplot(path = system.file("shiny", "microhaplot", package =
  "microhaplot"))
}
\arguments{
\item{path}{Path to shiny microhaplot app. Optional. If not specified, the path is default to local app path.}
}
\description{
Run shiny microhaplot app
}
\examples{
# this starts a shiny session, so should not be run during R CMD CHECK, etc.
\dontrun{
runShinyHaplot()
}
}
