% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMicroHap.R
\name{RunSrMicrohap}
\alias{RunSrMicrohap}
\title{Microhaplotype sampler}
\usage{
RunSrMicrohap(haplo.tbl, locus, n.sam, random.seed = 43454,
  prior.model = "uniform", min.read.depth = 0)
}
\arguments{
\item{haplo.tbl}{data frame. A haplotype data frame generated from running `runHaplo`.  The table contains "group" label, individual "id" label,
"locus" label, observed "haplo"type, haplotype read "depth", "logP.call", "logP.miscall", "pos", "allele.balance","rank", etc.}

\item{locus}{string. the locus name. Required}

\item{n.sam}{integer. number of iterations. required}

\item{random.seed}{integer. Set the random seed number. Default sets as 43454. Optional}

\item{prior.model}{String. Choose two different prior models: "uniform" - prior set all prior haplotype weight to 1, or "empirical": the prior
alpha values are defined by the number of observed cases under refinement}

\item{min.read.depth}{integer. Set minimal read depth. Default as 0, as no minimal.}
}
\description{
This hierarchical Gibbs sampler infers the true haplotype of grouped individuals of a given locus.
}
\examples{
# collect.data <- RunSrMicrohap(haplo.sum, "tag_id_1377", 2)
}
