% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplot.R
\name{mvShinyHaplot}
\alias{mvShinyHaplot}
\title{Transfer a copy of microhaplot app.}
\usage{
mvShinyHaplot(path)
}
\arguments{
\item{path}{string. directory path. Required}
}
\value{
a logical value of whether \code{file.copy} is successfully transferred Shiny app to its new directory
}
\description{
Moves shiny microhaplot app to a different directory
}
\examples{

mvShinyHaplot(tempdir())

}
