% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplot.R
\name{runHaplotype}
\alias{runHaplotype}
\title{Run Haplotype Plot.}
\usage{
runHaplotype(path = system.file("shiny", "haPLOType", package = "haplot"))
}
\description{
Run Haplotype shiny
}
\examples{
runHaplotype()
}

