% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplot.R
\name{runShinyHaplot}
\alias{runShinyHaplot}
\title{Run shiny microhaplot}
\usage{
runShinyHaplot(path = system.file("shiny", "microhaplot", package =
  "microhaplot"))
}
\arguments{
\item{path}{Path to shiny microhaplot app. Optional. If not specified, the path is default to local app path.}
}
\value{
Runs shiny microhaplot application via \code{shiny::runApp} which typically doesn't return; interrupt R to stop the application (usually by pressing Ctrl+C or Esc).
}
\description{
Run shiny microhaplot app
}
\examples{
if(interactive()){
runShinyHaplot()
}
}
